# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Provisum API
  version: 0.7.0
  description: |
    REST API for Provisum — enterprise security role mapping platform for ERP migrations
    (e.g. SAP ECC to S/4HANA). Manages the full workflow: data upload, AI persona generation,
    role mapping, SOD conflict analysis, approvals, and provisioning exports.

    **Authentication:** All endpoints (except `/api/auth/login`, `/api/auth/setup`, and
    `/api/health`) require a valid Supabase JWT session cookie. The cookie is set automatically
    on successful login via `/api/auth/login`.

    **Role hierarchy (higher = more access):**
    `system_admin (100) > admin (80) > approver (60) > coordinator (50) > mapper (40) > viewer (20)`
  contact:
    name: Provisum Support
    url: https://provisum.io
  license:
    name: Proprietary
servers:
  - url: https://demo.provisum.io
    description: Demo instance
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Supabase JWT session cookie (set by /api/auth/login)
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "Unauthorized"
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
...
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Unauthorized"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Insufficient permissions"
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
security:
  - cookieAuth: []
paths:
...
  /api/admin/feature-flags:
    get:
      tags: [Admin]
      summary: Get feature flags
      description: Returns the current state of all feature flags. Requires admin or system_admin.
      responses:
        "200":
          description: Feature flags
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: boolean
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Admin]
      summary: Update feature flags
      description: Toggles one or more feature flags. Requires admin or system_admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: boolean
      responses:
        "200":
          description: Flags updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/ServerError"

  # ===========================================================================
  # Admin - Users (Invitations)
  # ===========================================================================
...
  /api/admin/webhooks:
    get:
      tags: [Admin]
      summary: List webhook endpoints or deliveries
      description: |
        Returns all webhook endpoints and supported event types. If `endpointId` query param
        is provided, returns delivery history for that endpoint. Requires admin or system_admin.
      parameters:
        - name: endpointId
          in: query
          required: false
          schema:
            type: integer
          description: If provided, returns deliveries for this endpoint instead of the endpoint list
      responses:
        "200":
          description: Webhook endpoints or deliveries
          content:
            application/json:
              schema:
                type: object
                properties:
                  endpoints:
                    type: array
                    items:
                      type: object
                  eventTypes:
                    type: array
                    items:
                      type: string
                  deliveries:
                    type: array
                    items:
                      type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
    post:
      tags: [Admin]
      summary: Create a webhook endpoint
      description: Creates a new webhook endpoint with a generated signing secret. Requires admin or system_admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, events]
              properties:
                url:
                  type: string
                  format: uri
                  description: Must be an HTTPS URL
                description:
                  type: string
                events:
                  type: array
                  items:
                    type: string
                  description: Event types to subscribe to
                enabled:
                  type: boolean
                  default: true
      responses:
        "200":
          description: Webhook created (includes signing secret)
          content:
            application/json:
              schema:
                type: object
                properties:
                  endpoint:
                    type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
    patch:
      tags: [Admin]
      summary: Update a webhook endpoint
      description: Updates fields on an existing webhook endpoint. Requires admin or system_admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id:
                  type: integer
                url:
                  type: string
                description:
                  type: string
                events:
                  type: array
                  items:
                    type: string
                enabled:
                  type: boolean
      responses:
        "200":
          description: Webhook updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      tags: [Admin]
      summary: Delete a webhook endpoint
      description: Deletes a webhook endpoint and all its delivery history. Requires admin or system_admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id:
                  type: integer
      responses:
        "200":
          description: Webhook deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"

  # ===========================================================================
  # Admin - Scheduled Exports
  # ===========================================================================
  /api/admin/scheduled-exports:
    get:
      tags: [Admin]
      summary: List scheduled exports
      description: Returns all scheduled export configurations and supported export types. Requires admin or system_admin.
      responses:
        "200":
          description: Scheduled exports list
          content:
            application/json:
              schema:
                type: object
                properties:
                  exports:
                    type: array
                    items:
                      type: object
                  exportTypes:
                    type: array
                    items:
                      type: object
                      properties:
                        value:
                          type: string
                        label:
                          type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
    post:
      tags: [Admin]
      summary: Create a scheduled export
      description: Creates a new recurring export schedule. Requires admin or system_admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, exportType, schedule]
              properties:
                name:
                  type: string
                exportType:
                  type: string
                schedule:
                  type: string
                  enum: [daily, weekly, monthly]
                dayOfWeek:
                  type: integer
                  nullable: true
                  description: 0-6 (Sunday-Saturday), used for weekly
                dayOfMonth:
                  type: integer
                  nullable: true
                  description: 1-31, used for monthly
                hour:
                  type: integer
                  default: 6
                  description: Hour of day (0-23) to run the export
                enabled:
                  type: boolean
                  default: true
      responses:
        "200":
          description: Scheduled export created
          content:
            application/json:
              schema:
                type: object
                properties:
                  export:
                    type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
    patch:
      tags: [Admin]
      summary: Update a scheduled export
      description: Updates fields on an existing scheduled export. Recalculates next run time if schedule changes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id:
                  type: integer
                name:
                  type: string
                schedule:
                  type: string
                  enum: [daily, weekly, monthly]
                dayOfWeek:
                  type: integer
                  nullable: true
                dayOfMonth:
                  type: integer
                  nullable: true
                hour:
                  type: integer
                enabled:
                  type: boolean
      responses:
        "200":
          description: Scheduled export updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      tags: [Admin]
      summary: Delete a scheduled export
      description: Permanently deletes a scheduled export configuration. Requires admin or system_admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id:
                  type: integer
      responses:
        "200":
          description: Scheduled export deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"

  # ===========================================================================
  # Calibration
  # ===========================================================================
...